Add python ITK based calculation of GLCM
itk_functions: ITK based functions to calculate the GLCM. Running this script compares the output of 2 ITK based methods (either with or without intermediate output of the GLCM matrices) and PyRadiomics (with alternative binning enabled). Computed feature values should match up
compute_fmap: Defines a simple iterator over the mask and uses it to extract an image region, from which GLCM features are calculated using ITK. This represents voxel-based feature calculation using full-kernels and kernel-level binning (ITK method). The output can be compared against the output of per_kernel_normalization as it is setup in this commit.